lots of cleanup in opengl renderer
[fixed] most animations.
[fixed] uuid bug with serialization
[fixed] double precision error
[fixed] shaders taking a vec4 when they should've been taking vec3
[fixed] cache model matrix bug, removed cache for now
[fixed] uint32_t for all index types related to bones

[added] uniform directional light so its easier to see white on white
[added] speed/distance to line traversal
[added] look direction to object when moving along line
[added] test walk cycles

+++ Core/CubicSpline.cpp
@@ -10,7 +10,7 @@
 
 #include "CubicSpline.h"
 
-static float TruncatedPower(float t, float c, float d)
+static double TruncatedPower(double t, double c, double d)
 {
     if (t < c) return 0;
 
@@ -28,11 +28,11 @@ namespace Gep
         ComputeCoeffs();
     }
 
-    glm::vec3 CubicSpline::Evaluate(float t) const
+    glm::vec3 CubicSpline::Evaluate(double t) const
     {
         if (controlPointCount < 2) return glm::vec3(0.0f);
 
-        t *= static_cast<float>(controlPointCount - 1); // normalize the t value
+        t *= static_cast<double>(controlPointCount - 1); // normalize the t value
 
         glm::vec3 vt{
             coeffsx[0] + coeffsx[1] * t + coeffsx[2] * t * t + coeffsx[3] * t * t * t,
@@ -42,7 +42,7 @@ namespace Gep
 
         for (uint32_t j = 1; j <= controlPointCount - 2; ++j)
         {
-            const float tp = TruncatedPower(t, static_cast<float>(j), 3.0f);
+            const double tp = TruncatedPower(t, static_cast<double>(j), 3.0f);
             vt.x += coeffsx[j + 3] * tp;
             vt.y += coeffsy[j + 3] * tp;
             vt.z += coeffsz[j + 3] * tp;
@@ -72,10 +72,10 @@ namespace Gep
         // Data fitting rows (t = 0,1,...,k)
         for (uint32_t i = 0; i < n; ++i)
         {
-            const float t = static_cast<float>(i);
+            const double t = static_cast<double>(i);
 
             // polynomial part: [1, t, t^2, t^3]
-            A(i, 0) = 1.0f;
+            A(i, 0) = 1.0;
             A(i, 1) = t;
             A(i, 2) = t * t;
             A(i, 3) = t * t * t;
@@ -83,7 +83,7 @@ namespace Gep
             // truncated power terms: (t - j)^3_+ for j = 1..k-1
             for (uint32_t j = 1; j <= k - 1; ++j)
             {
-                A(i, j + 3) = TruncatedPower(t, static_cast<float>(j), 3.0f);
+                A(i, j + 3) = TruncatedPower(t, static_cast<double>(j), 3.0);
             }
 
             bx[i] = controlPoints[i].x;
@@ -97,18 +97,18 @@ namespace Gep
 
         // s''(t) = 2c + 6dt + sum_j 6(t - j)_+ * alpha_j
         // At t = 0: 2c = 0
-        A(r0, 2) = 2.0f;
+        A(r0, 2) = 2.0;
 
         // At t = k: 2c + 6dk + sum_j 6(k - j) * alpha_j = 0
-        A(r1, 3) = 6.0f * static_cast<float>(k);
+        A(r1, 3) = 6.0 * static_cast<double>(k);
         for (uint32_t j = 1; j <= k - 1; ++j)
         {
-            A(r1, j + 3) = 6.0f * static_cast<float>(k - j);
+            A(r1, j + 3) = 6.0 * static_cast<double>(k - j);
         }
 
         // RHS for boundary rows (zeros) — previously missing/incorrect
-        bx[r0] = by[r0] = bz[r0] = 0.0f;
-        bx[r1] = by[r1] = bz[r1] = 0.0f;
+        bx[r0] = by[r0] = bz[r0] = 0.0;
+        bx[r1] = by[r1] = bz[r1] = 0.0;
     }
 
     void CubicSpline::ComputeCoeffs()
